refactor(tx-builder): move buildTxAsync to builder folder
