+++ src/contract/Contract.ts
@@ -25,8 +25,9 @@
 import { Encoder as Calldata } from '@aeternity/aepp-calldata';
 import { DRY_RUN_ACCOUNT } from '../tx/builder/schema';
 import { Tag, AensName } from '../tx/builder/constants';
-import { buildContractIdByContractTx, unpackTx } from '../tx/builder';
-import { _buildTx, BuildTxOptions } from '../tx';
+import {
+  buildContractIdByContractTx, unpackTx, buildTxAsync, BuildTxOptions,
+} from '../tx/builder';
 import { send, SendOptions } from '../spend';
 import { decode, Encoded } from '../utils/encoder';
 import {
@@ -280,7 +281,7 @@ class Contract<M extends ContractMethodsBase> {
     if (opt.onAccount == null) throw new IllegalArgumentError('Can\'t deploy without account');
     const ownerId = opt.onAccount.address;
     if (this.$options.bytecode == null) throw new IllegalArgumentError('Can\'t deploy without bytecode');
-    const tx = await _buildTx({
+    const tx = await buildTxAsync({
       ...opt,
       tag: Tag.ContractCreateTx,
       gasLimit: opt.gasLimit ?? await this._estimateGas('init', params, opt),
@@ -371,12 +372,12 @@ class Contract<M extends ContractMethodsBase> {
       let tx;
       if (fn === 'init') {
         if (this.$options.bytecode == null) throw new IllegalArgumentError('Can\'t dry-run "init" without bytecode');
-        tx = await _buildTx({
+        tx = await buildTxAsync({
           ...txOpt, tag: Tag.ContractCreateTx, code: this.$options.bytecode, ownerId: callerId,
         });
       } else {
         if (contractId == null) throw new MissingContractAddressError('Can\'t dry-run contract without address');
-        tx = await _buildTx({
+        tx = await buildTxAsync({
           ...txOpt, tag: Tag.ContractCallTx, callerId, contractId,
         });
       }
@@ -391,7 +392,7 @@ class Contract<M extends ContractMethodsBase> {
     } else {
       if (top != null) throw new IllegalArgumentError('Can\'t handle `top` option in on-chain contract call');
       if (contractId == null) throw new MissingContractAddressError('Can\'t call contract without address');
-      const tx = await _buildTx({
+      const tx = await buildTxAsync({
         ...opt,
         tag: Tag.ContractCallTx,
         gasLimit: opt.gasLimit ?? await this._estimateGas(fn, params, opt),
